+++ common/src/main/java/net/spell_engine/rpg_series/datagen/WeaponSkills.java
@@ -386,7 +386,7 @@ public class WeaponSkills {
         attack.hitbox.arc = 160;
         attack.hitbox.height = 0.2F;
         attack.hitbox.width = 0.5F;
-        attack.forward_momentum = 1.5F;
+        attack.forward_momentum = 2.25F;
         attack.movement_slipperiness = 0.25F;
 
         attack.additional_strikes = 4;
@@ -421,7 +421,7 @@ public class WeaponSkills {
         attack.hitbox.arc = 160;
         attack.hitbox.length = 0.5F;
         attack.hitbox.height = 0.2F;
-        attack.forward_momentum = 2F;
+        attack.forward_momentum = 1.5F;
         attack.movement_slipperiness = 0.3F;
         attack.delay = 0.1F;
         attack.additional_strikes = 4;